requirements.txt
asgiref==3.2.7
Django==3.0.7
django-ckeditor==5.9.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-js-asset==1.2.2
djangorestframework==3.11.0
Markdown==3.2.2
mysqlclient==1.4.6
pytz==2020.1
sqlparse==0.3.1
